Added gitignore

+++ management/models.py
@@ -19,5 +19,4 @@ class Transaction(models.Model):
     amount = models.IntegerField()
     date = models.DateField()
     feedback = models.TextField()
-    sentiment=models.IntegerField(default=0)
-    
+    sentiment = models.FloatField(default=0.0)
